Minor refactoring: Restructure the code environment; remove 'TranslationMemory' namespace
