+++ Sdl.Community.Toolkit.LanguagePlatform/EnvironmentConstants.cs
@@ -1,4 +1,4 @@
-﻿namespace Sdl.Community.Toolkit.LanguagePlatform.TranslationMemory
+﻿namespace Sdl.Community.Toolkit.LanguagePlatform
 {
 	/// <summary>
 	///
