+++ Sdl.Community.Toolkit.LanguagePlatform/Visitors/SegmentVisitor.cs
@@ -11,10 +11,10 @@ namespace Sdl.Community.Toolkit.LanguagePlatform.Visitors
 		private int TagId { get; set; }
 
 		public bool HasRevisions { get; private set; }
-		public Sdl.LanguagePlatform.Core.Segment Segment { get; }
+		public Segment Segment { get; }
 		public List<IComment> Comments { get; set; }
 
-		public SegmentVisitor(Sdl.LanguagePlatform.Core.Segment segment, bool ignoreTags)
+		public SegmentVisitor(Segment segment, bool ignoreTags)
 		{
 			Segment = segment;
 			IgnoreTags = ignoreTags;
